refactor: retire unilab.config package
structured_configs promoted to top-level unilab.structured_configs;
reward helpers folded into unilab.training.reward. The two *_params
modules held only unused KNOWN_TASKS stubs and are deleted.
Eliminates the config/ vs conf/ naming collision at the source.
