--- src/unilab/config/__init__.py
@@ -1,3 +1,0 @@
-from unilab.config.reward import RewardDict, extract_reward_config, resolve_reward_dict
-
-__all__ = ["RewardDict", "extract_reward_config", "resolve_reward_dict"]
