--- src/unilab/config/locomotion_params.py
@@ -1,16 +1,0 @@
-"""Locomotion task registry.
-
-Config factory functions (ppo_config, appo_config, offpolicy_config) have been
-removed. Configurations are now managed via Hydra YAML files in conf/.
-KNOWN_TASKS is kept for legacy routing utilities.
-"""
-
-KNOWN_TASKS: frozenset[str] = frozenset(
-    {
-        "Go1JoystickFlat",
-        "Go2JoystickFlat",
-        "G1WalkFlat",
-        "G1WalkRough",
-        "G1MotionTracking",
-    }
-)
